feat(oauth): adding discord oauth support
Initial pass here, right now the discord oauth token is being saved
entirely on the client side. I'm thinking it would be better to store this
token on the serverside and have a session saved on the client that would
link back to this. Some other refactors in here aswell.